refactor(SDK-1119): nest onboarding Contractor Profile under onboarding/
Aligns Contractor/Profile with the Employee domain's onboarding/management/shared
convention (and Contractor's own ContractorList precedent), where the older
onboarding component previously sat at the domain root instead of alongside its
new management/shared siblings. Pure move; no behavior change.
