--- src/components/Contractor/Profile/index.ts
@@ -1,6 +1,0 @@
-export {
-  ContractorProfile,
-  type ContractorProfileProps,
-  type ContractorProfileAdminProps,
-  type ContractorProfileSelfOnboardingProps,
-} from './ContractorProfile'
