+++ src/components/Contractor/Profile/onboarding/IndividualSelfOnboardingProfile.tsx
@@ -3,27 +3,27 @@ import classNames from 'classnames'
 import {
   useContractorDetailsForm,
   type ContractorDetailsOptionalFieldsToRequire,
-} from './shared/useContractorDetailsForm'
+} from '../shared/useContractorDetailsForm'
 import type { ContractorSelfOnboardingProfileProps } from './SelfOnboardingContractorProfile'
 import styles from './ContractorProfile.module.scss'
 import { BaseLayout } from '@/components/Base'
 import { SDKFormProvider } from '@/partner-hook-utils/form/SDKFormProvider'
 import { Form } from '@/components/Common/Form'
-import { Flex } from '@/components/Common'
+import { Flex, Grid } from '@/components/Common'
 import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
 import { componentEvents } from '@/shared/constants'
 
-const BUSINESS_REQUIRED_FIELDS: ContractorDetailsOptionalFieldsToRequire = {
-  update: ['businessName', 'ein'],
+const INDIVIDUAL_REQUIRED_FIELDS: ContractorDetailsOptionalFieldsToRequire = {
+  update: ['firstName', 'lastName', 'ssn'],
 }
 
 /**
- * Self-onboarding profile for a business contractor — collects business name
- * and EIN.
+ * Self-onboarding profile for an individual contractor — collects legal name
+ * and SSN.
  *
  * @internal
  */
-export function BusinessSelfOnboardingProfile({
+export function IndividualSelfOnboardingProfile({
   contractorId,
   onEvent,
   className,
@@ -35,7 +35,7 @@ export function BusinessSelfOnboardingProfile({
     contractorId,
     withSelfOnboardingField: false,
     defaultValues: { selfOnboarding: true },
-    optionalFieldsToRequire: BUSINESS_REQUIRED_FIELDS,
+    optionalFieldsToRequire: INDIVIDUAL_REQUIRED_FIELDS,
   })
 
   if (contractor.isLoading) {
@@ -66,23 +66,38 @@ export function BusinessSelfOnboardingProfile({
                 <Flex flexDirection="column" gap={4}>
                   <Components.Heading as="h2">{t('selfOnboarding.title')}</Components.Heading>
                   <Components.Text variant="supporting">
-                    {t('selfOnboarding.businessDescription')}
+                    {t('selfOnboarding.individualDescription')}
                   </Components.Text>
                 </Flex>
               </header>
 
-              {Fields.BusinessName && (
+              {Fields.FirstName && Fields.LastName && (
                 <>
-                  <Fields.BusinessName
-                    label={t('fields.businessName.label')}
-                    validationMessages={{ REQUIRED: t('validations.businessName') }}
-                  />
-                  {Fields.Ein && (
-                    <Fields.Ein
-                      label={t('fields.ein.label')}
+                  <Grid gridTemplateColumns={{ base: '1fr', medium: '1fr 1fr' }} gap={16}>
+                    <Fields.FirstName
+                      label={t('fields.firstName.label')}
                       validationMessages={{
-                        INVALID_EIN: t('validations.einFormat'),
-                        REQUIRED: t('validations.ein'),
+                        REQUIRED: t('validations.firstName'),
+                        INVALID_NAME: t('validations.firstNameFormat'),
+                      }}
+                    />
+                    {Fields.MiddleInitial && (
+                      <Fields.MiddleInitial label={t('fields.middleInitial.label')} />
+                    )}
+                  </Grid>
+                  <Fields.LastName
+                    label={t('fields.lastName.label')}
+                    validationMessages={{
+                      REQUIRED: t('validations.lastName'),
+                      INVALID_NAME: t('validations.lastNameFormat'),
+                    }}
+                  />
+                  {Fields.Ssn && (
+                    <Fields.Ssn
+                      label={t('fields.ssn.label')}
+                      validationMessages={{
+                        INVALID_SSN: t('validations.ssnFormat'),
+                        REQUIRED: t('validations.ssn'),
                       }}
                     />
                   )}
