+++ src/components/Contractor/SelfOnboardingFlow/SelfOnboardingComponents.tsx
@@ -3,7 +3,7 @@ import type { FlowContextInterface } from '@/components/Flow/useFlow'
 import { useFlow } from '@/components/Flow/useFlow'
 import type { BaseComponentInterface } from '@/components/Base'
 import { Landing as LandingComponent } from '@/components/Contractor/Landing'
-import { ContractorProfile as ProfileComponent } from '@/components/Contractor/Profile'
+import { ContractorProfile as ProfileComponent } from '@/components/Contractor/Profile/onboarding/ContractorProfile'
 import { Address as AddressComponent } from '@/components/Contractor/Address'
 import { PaymentMethod as PaymentMethodComponent } from '@/components/Contractor/PaymentMethod/PaymentMethod'
 import { DocumentSigner as DocumentSignerComponent } from '@/components/Contractor/Documents/DocumentSigner'
